Make targets and job to build plugin binaries
The build target builds the binary for the current OS/arch. 

The all target builds binaries for all targets in BUILD_ARCHS and
BUILD_PLATFORMS using gox. The BUILD_ARCHS and BUILD_PLATFORMS values
are straight from the gitlab-runner's Makefile . We can remove some if
they aren't relevant.

The job uses the default golang image for the golang version defined in
the common CI configuration (1.18.9), to stay in sync with runner.