Make target, script and job to upload binaries
Use the GitLab API to upload binaries to the generic package repo. The
resulting URL for the packages will be, for e.g.

https://gitlab.com/api/v4/projects/39457165/packages/generic/releases/v0.0.1/fleeting-plugin-aws-linux-amd64

Note that we pipe the list of binaries to a file because we need them
later when creating a release to link the release to the binary
artifacts.