Make target, script and job to make release
Following the direction in
https://gitlab.com/gitlab-org/release-cli/-/tree/master/docs/examples/release-assets-as-generic-package/,
this creates a release and links artifacts in the generic package
registry to the release.

Note that for the `latest` release, we have to delete the existing one
and create a new new since it's not possible to update an existing
release to point it to different artifacts. However, the `latest`
artifacts for all previous version will still exists in the generic
package registry.