(provider_test) Remove capacity sync test
This is relevant for the following providers:

- AWS: https://gitlab.com/gitlab-org/fleeting/fleeting-plugin-aws/-/blob/b73a837ad7c69f93e4919711bfc462cefbb2589e/provider_test.go
- Azure: https://gitlab.com/gitlab-org/fleeting/fleeting-plugin-azure/-/blob/f8ddf20a9c0a09d4b9f381cd96496bf7258512d0/provider_integration_test.go

...but in our case, there is simply no way to "increase the capacity" of
the instance group because the instance group doesn't formally "exist" in
the Hetzner world; it's an imaginary concept that we emulate by means of
labels set on the machines we create. Hence, we can safely remove this
for now.

MR: https://gitlab.com/hiboxsystems/fleeting-plugin-hetzner/-/merge_requests/5